Move code from inputs/psi to inputs/kernel
