--- plugins/inputs/all/psi.go
@@ -1,5 +1,0 @@
-//go:build !custom || inputs || inputs.psi
-
-package all
-
-import _ "github.com/influxdata/telegraf/plugins/inputs/psi" // register plugin
