--- plugins/inputs/psi/sample.conf
@@ -1,4 +1,0 @@
-# Read metrics about Pressure Stall Information (PSI)
-# Requires Linux kernel v4.20+
-[[inputs.psi]]
-  # no configuration
